requirements.txt
streamlit 
google.generativeai
langchain 
faiss-cpu 
pypdf 
tiktoken
